Create temp_workspace_script.py
This is the new script I'm working on to fit the needs of the new idea. I left some TODOs that we need to work on before we can start testing it. could use some help one the TODOs it'd just make things go by faster if you had an idea for it and just updated this file with that. Appreciate it Guys 😄